Fixes for IBATIS-414 (Abator) - allow generated example and criteria classes to be extended
git-svn-id: https://svn.apache.org/repos/asf/ibatis/trunk@544712 13f79535-47bb-0310-9956-ffa450edef68

+++ java/mapper/mapper2/tools/abator/core/testJava2/abatortest/java2/execute/conditional/java2/ConditionalJava2Tests.java
@@ -74,7 +74,7 @@ public class ConditionalJava2Tests extends BaseTest {
         try {
             Fieldsonly record = new Fieldsonly();
             record.setDoublefield(new Double(11.22));
-            record.setFloatfield(new Float(33.44));
+            record.setFloatfield(new Double(33.44));
             record.setIntegerfield(new Integer(5));
             dao.insert(record);
 
@@ -101,19 +101,19 @@ public class ConditionalJava2Tests extends BaseTest {
         try {
             Fieldsonly record = new Fieldsonly();
             record.setDoublefield(new Double(11.22));
-            record.setFloatfield(new Float(33.44));
+            record.setFloatfield(new Double(33.44));
             record.setIntegerfield(new Integer(5));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(44.55));
-            record.setFloatfield(new Float(66.77));
+            record.setFloatfield(new Double(66.77));
             record.setIntegerfield(new Integer(8));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(88.99));
-            record.setFloatfield(new Float(100.111));
+            record.setFloatfield(new Double(100.111));
             record.setIntegerfield(new Integer(9));
             dao.insert(record);
 
@@ -137,19 +137,19 @@ public class ConditionalJava2Tests extends BaseTest {
         try {
             Fieldsonly record = new Fieldsonly();
             record.setDoublefield(new Double(11.22));
-            record.setFloatfield(new Float(33.44));
+            record.setFloatfield(new Double(33.44));
             record.setIntegerfield(new Integer(5));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(44.55));
-            record.setFloatfield(new Float(66.77));
+            record.setFloatfield(new Double(66.77));
             record.setIntegerfield(new Integer(8));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(88.99));
-            record.setFloatfield(new Float(100.111));
+            record.setFloatfield(new Double(100.111));
             record.setIntegerfield(new Integer(9));
             dao.insert(record);
 
@@ -169,19 +169,19 @@ public class ConditionalJava2Tests extends BaseTest {
         try {
             Fieldsonly record = new Fieldsonly();
             record.setDoublefield(new Double(11.22));
-            record.setFloatfield(new Float(33.44));
+            record.setFloatfield(new Double(33.44));
             record.setIntegerfield(new Integer(5));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(44.55));
-            record.setFloatfield(new Float(66.77));
+            record.setFloatfield(new Double(66.77));
             record.setIntegerfield(new Integer(8));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(88.99));
-            record.setFloatfield(new Float(100.111));
+            record.setFloatfield(new Double(100.111));
             record.setIntegerfield(new Integer(9));
             dao.insert(record);
