+++ java/mapper/mapper2/tools/abator/core/testJava2/abatortest/java2/execute/flat/java2/FlatJava2Tests.java
@@ -70,7 +70,7 @@ public class FlatJava2Tests extends BaseTest {
         try {
             Fieldsonly record = new Fieldsonly();
             record.setDoublefield(new Double(11.22));
-            record.setFloatfield(new Float(33.44));
+            record.setFloatfield(new Double(33.44));
             record.setIntegerfield(new Integer(5));
             dao.insert(record);
 
@@ -97,19 +97,19 @@ public class FlatJava2Tests extends BaseTest {
         try {
             Fieldsonly record = new Fieldsonly();
             record.setDoublefield(new Double(11.22));
-            record.setFloatfield(new Float(33.44));
+            record.setFloatfield(new Double(33.44));
             record.setIntegerfield(new Integer(5));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(44.55));
-            record.setFloatfield(new Float(66.77));
+            record.setFloatfield(new Double(66.77));
             record.setIntegerfield(new Integer(8));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(88.99));
-            record.setFloatfield(new Float(100.111));
+            record.setFloatfield(new Double(100.111));
             record.setIntegerfield(new Integer(9));
             dao.insert(record);
 
@@ -133,19 +133,19 @@ public class FlatJava2Tests extends BaseTest {
         try {
             Fieldsonly record = new Fieldsonly();
             record.setDoublefield(new Double(11.22));
-            record.setFloatfield(new Float(33.44));
+            record.setFloatfield(new Double(33.44));
             record.setIntegerfield(new Integer(5));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(44.55));
-            record.setFloatfield(new Float(66.77));
+            record.setFloatfield(new Double(66.77));
             record.setIntegerfield(new Integer(8));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(88.99));
-            record.setFloatfield(new Float(100.111));
+            record.setFloatfield(new Double(100.111));
             record.setIntegerfield(new Integer(9));
             dao.insert(record);
 
@@ -165,19 +165,19 @@ public class FlatJava2Tests extends BaseTest {
         try {
             Fieldsonly record = new Fieldsonly();
             record.setDoublefield(new Double(11.22));
-            record.setFloatfield(new Float(33.44));
+            record.setFloatfield(new Double(33.44));
             record.setIntegerfield(new Integer(5));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(44.55));
-            record.setFloatfield(new Float(66.77));
+            record.setFloatfield(new Double(66.77));
             record.setIntegerfield(new Integer(8));
             dao.insert(record);
 
             record = new Fieldsonly();
             record.setDoublefield(new Double(88.99));
-            record.setFloatfield(new Float(100.111));
+            record.setFloatfield(new Double(100.111));
             record.setIntegerfield(new Integer(9));
             dao.insert(record);
